🐳chore(data): add game_info data
This data named `game_info` was previously living in `tarflow.iquizoo` package. But it is more reasonable to put it here. The data `game_info` contains the configurations of the required preprocessing function names of all the games, and this package contains all the functions which makes a configuration of that very natural.

Signed-off-by: Liang Zhang <psychelzh@outlook.com>